feat: add v2 API models and interactions, including email handling and person merge functionality
