+++ affinity_model/.gitattributes
@@ -1,1 +1,1 @@
-v2.py linguist-generated=true
+v2/ linguist-generated=true
