+++ affinity_model/base.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class MyBaseModel(BaseModel):
+    pass
